Move worker thread into separate module and start before UI
