+++ crates/hulk_replayer/src/window.rs
@@ -1,22 +1,17 @@
-use std::{
-    collections::BTreeMap,
-    thread::spawn,
-    time::{Duration, SystemTime},
-};
+use std::{collections::BTreeMap, time::SystemTime};
 
 use eframe::{
     egui::{CentralPanel, Context},
-    App, CreationContext, Frame,
+    App, Frame,
 };
+use tokio::sync::watch;
+
 use framework::Timing;
-use tokio::{runtime::Builder, select, sync::watch, time::sleep};
 
 use crate::{
     coordinate_systems::{AbsoluteTime, FrameRange, RelativeTime, ViewportRange},
-    execution::Replayer,
     labels::Labels,
     timeline::Timeline,
-    ReplayerHardwareInterface,
 };
 
 pub struct Window {
@@ -29,19 +24,11 @@ pub struct Window {
 
 impl Window {
     pub fn new(
-        creation_context: &CreationContext,
-        replayer: Replayer<ReplayerHardwareInterface>,
+        indices: BTreeMap<String, Vec<Timing>>,
+        time_sender: watch::Sender<SystemTime>,
     ) -> Self {
-        let frame_range = join_timing(&replayer);
+        let frame_range = join_timing(&indices);
         let viewport_range = ViewportRange::from_frame_range(&frame_range);
-
-        let (time_sender, time_receiver) = watch::channel(SystemTime::UNIX_EPOCH);
-        let indices = replayer
-            .get_recording_indices()
-            .into_iter()
-            .map(|(name, index)| (name, index.iter().collect()))
-            .collect();
-        spawn_replay_thread(replayer, creation_context.egui_ctx.clone(), time_receiver);
 
         Self {
             time_sender,
@@ -81,52 +68,20 @@ impl App for Window {
     }
 }
 
-fn join_timing(replayer: &Replayer<ReplayerHardwareInterface>) -> FrameRange {
-    let recording_indices = replayer.get_recording_indices();
-    let begin = recording_indices
+fn join_timing(indices: &BTreeMap<String, Vec<Timing>>) -> FrameRange {
+    let begin = indices
         .values()
-        .flat_map(|index| index.first_timing().map(|timing| timing.timestamp))
+        .flat_map(|index| index.first().map(|timing| timing.timestamp))
         .min()
         .unwrap_or(SystemTime::UNIX_EPOCH);
-    let end = recording_indices
+    let end = indices
         .values()
         .flat_map(|index| {
             index
-                .last_timing()
+                .last()
                 .map(|timing| timing.timestamp + timing.duration)
         })
         .max()
         .unwrap_or(SystemTime::UNIX_EPOCH);
     FrameRange::new(AbsoluteTime::new(begin), AbsoluteTime::new(end))
-}
-
-fn spawn_replay_thread(
-    mut replayer: Replayer<ReplayerHardwareInterface>,
-    egui_context: Context,
-    mut time: watch::Receiver<SystemTime>,
-) {
-    spawn(move || {
-        let runtime = Builder::new_current_thread().enable_all().build().unwrap();
-
-        runtime.block_on(async move {
-            let parameters_changed = replayer.get_parameters_changed();
-            loop {
-                select! {
-                    _ = parameters_changed.notified() => {}
-                    _ = sleep(Duration::from_secs(1)) => {}
-                    result = time.changed() => {
-                        if result.is_err() {
-                            // channel closed, quit thread
-                            break;
-                        }
-                    }
-                }
-
-                if let Err(error) = replayer.replay_at(*time.borrow()) {
-                    eprintln!("{error}");
-                }
-                egui_context.request_repaint();
-            }
-        });
-    });
 }
